add function to create mainObj and fix date formatting
In order to analyse data using the dates, a consistent format for the dates is needed. Added function to getDateByYearOnly which can be applied to the data object. It adds an array to the object to allow for checking integrity with the original dataset. Having done this the entire data object can be sorted by Date without losing tract of individual properties eg HULLDESC, COUNTRY, RIGDESC etc